// src/app/loans/models/loan-product.model.ts
/**
 * Copyright since 2025 Mifos Initiative
 *
 * This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/.
 */

import { Currency } from 'app/shared/models/general.model';

export interface LoanProductBasicDetails {
  productType: string;
  id: number;
  name: string;
  shortName: string;
  description: string;
  currency: Currency;
}